requirements.txt
flask==3.0.2
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
Pillow==10.2.0
scikit-learn==1.4.1.post1
gunicorn==21.2.0